% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_classification_tbl.R
\name{get_classification_tbl}
\alias{get_classification_tbl}
\title{Create a tibble of classifications of all functions used in your analysis}
\usage{
get_classification_tbl(func_names = NULL)
}
\arguments{
\item{func_names}{Character vector of the names of all functions used in analysis. If `NULL`,
will return the full data frame of all classified functions.}
}
\value{
Tibble with two columns, `classification` and `func`
}
\description{
Create a tibble of classifications of all functions used in your analysis
}
