% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_classifications.R
\name{get_classifications}
\alias{get_classifications}
\title{Get a tidy data frame of classifications of all functions used in your analysis}
\usage{
get_classifications(func_names = NULL, include_duplicates = TRUE)
}
\arguments{
\item{func_names}{Character vector of the names of all functions used in
analysis. If `NULL`, will return the full data frame of all classified functions.}

\item{include_duplicates}{Logical. Indicates whether to include all functions
and classifications along with their score (default, `TRUE`) - this may
result in multiple lines (with multiple classifications) for a single function.
If `FALSE`, the most prevalent classification will be selected.}
}
\value{
If `include_duplicates = TRUE`, tibble with four columns,
`classification` and `func`, `score`.
If `include_duplicates = FALSE`, tibble with two columns, `classification` and
`func`.
}
\description{
Get a tidy data frame of classifications of all functions used in your analysis
}
\examples{
# Get a data frame of all classifications
get_classifications()

# Get a data frame of the most prevalent classifications
get_classifications(include_duplicates = FALSE)
}
