% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_functions.R
\name{get_package_functions}
\alias{get_package_functions}
\title{Get Package Functions}
\usage{
get_package_functions(x)
}
\arguments{
\item{x}{an R call or list of R calls}
}
\value{
a data frame with the following columns:
\itemize{
\item \code{package}: The name of the package
\item \code{func}: The name of the function
}
}
\description{
Loads all packages called in a list of R calls and creates a data frame of
all functions included in these packages. Packages that were not previously
loaded that are loaded as part of this call will be detached.
}
\examples{
get_package_functions(
  list(
    quote(library(tidycode)),
    quote(library(purrr))
  )
)
}
