% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_tbl.R
\name{get_packages_tbl}
\alias{get_packages_tbl}
\title{Create a tibble of all functions included in all packages used in your analysis}
\usage{
get_packages_tbl(pkg_names = NULL)
}
\arguments{
\item{pkg_names}{Character vector of the names of all packages used in analysis}
}
\value{
A \code{tbl_df} with two columns:
\itemize{
\item \code{package}: the name of the R package
\item \code{func}: the name of the R function
}
}
\description{
Create a tibble of all functions included in all packages used in your analysis
}
