% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stopfuncs.R
\name{get_stopfuncs}
\alias{get_stopfuncs}
\title{Get a tidy data frame of a "stopword" lexicon for R functions}
\usage{
get_stopfuncs()
}
\value{
A tibble with one column, `func`.
}
\description{
Get a data frame listing one function per row.
}
\examples{
get_stopfuncs()
}
