% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_model.R
\name{is_model}
\alias{is_model}
\title{Check if value is of a model class}
\usage{
is_model(x)
}
\arguments{
\item{x}{value of R expression}
}
\value{
logical
}
\description{
Check if value is of a model class
}
\examples{
val <- matahari::dance_tbl()$value
purrr::map_lgl(val, is_model)
}
