% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_package.R
\name{is_package}
\alias{is_package}
\title{Check if an expression is loading an R package}
\usage{
is_package(x)
}
\arguments{
\item{x}{an R call or list of R calls}
}
\value{
logical value or list of logical values
}
\description{
Check if an expression is loading an R package
}
\examples{
is_package(quote(library(tidycode)))
is_package(quote(tidycode::is_package))
is_package(
  list(
    quote(library(tidycode)),
    quote(library(purrr)))
  )
}
