% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_pipe.R
\name{is_pipe}
\alias{is_pipe}
\title{Check if an expression includes a pipe}
\usage{
is_pipe(x)
}
\arguments{
\item{x}{an R call}
}
\value{
logical
}
\description{
Check if an expression includes a pipe
}
\examples{
matahari::dance_start()
mtcars$mpg \%>\% mean()
matahari::dance_stop()
expr <- matahari::dance_tbl()$expr
is_pipe(expr)
matahari::dance_remove()

}
