% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_packages.R
\name{load_packages}
\alias{load_packages}
\title{Load packages}
\usage{
load_packages(x)
}
\arguments{
\item{x}{an R call or list of R calls}
}
\description{
Loads all R packages in the list of R calls.
}
\examples{
load_packages(
  list(
    quote(library(tidycode)),
    quote(library(purrr))
  )
)
}
