% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_fun_args.R
\name{ls_fun_args}
\alias{ls_fun_args}
\title{List arguments in R call}
\usage{
ls_fun_args(x)
}
\arguments{
\item{x}{an R call}
}
\value{
List, all arguments in an R call
}
\description{
List arguments in R call
}
\examples{
ls_fun_args(quote(lm(mpg ~ cyl, mtcars)))
}
