% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_fun_calls.R
\name{ls_fun_calls}
\alias{ls_fun_calls}
\title{List functions in R call}
\usage{
ls_fun_calls(x)
}
\arguments{
\item{x}{an R call or list of R calls}
}
\value{
Character, all functions in an R call
}
\description{
List functions in R call
}
\examples{
ls_fun_calls(quote(lm(mpg ~ cyl, mtcars)))
}
