% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_models.R
\name{ls_models}
\alias{ls_models}
\title{List models}
\usage{
ls_models(x)
}
\arguments{
\item{x}{an R call or list of R calls}
}
\value{
Character. Vector of unique models called during matahari session.
}
\description{
List models
}
\examples{
matahari::dance_start()
m <- lm(mpg ~ cyl, mtcars)
matahari::dance_stop()
expr <- matahari::dance_tbl()$expr
ls_models(expr)
matahari::dance_remove()
}
