% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_pipe.R
\name{ls_pipes}
\alias{ls_pipes}
\title{List pipes}
\usage{
ls_pipes(x)
}
\arguments{
\item{x}{list of R calls}
}
\value{
Character. Vector of piped calls during matahari session.
}
\description{
List pipes
}
\examples{
matahari::dance_start()
mtcars$mpg \%>\% mean()
matahari::dance_stop()
expr <- matahari::dance_tbl()$expr
matahari::dance_remove()
}
