% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_plot.R
\name{ls_plots}
\alias{ls_plots}
\title{List plots}
\usage{
ls_plots(x)
}
\arguments{
\item{x}{an R call or list of R calls}
}
\value{
Character. Vector of plot calls during matahari session.
}
\description{
List plots
}
\examples{
matahari::dance_start()
plot(mtcars$mpg)
matahari::dance_stop()
expr <- matahari::dance_tbl()$expr
ls_plots(expr)
matahari::dance_remove()
}
