% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rfiles.R
\name{read_rfiles}
\alias{read_rfiles}
\title{Read R file(s) as a tidy data frame}
\usage{
read_rfiles(...)
}
\arguments{
\item{...}{One or more quoted R file paths to read}
}
\value{
A tidy data frame, a \code{tbl_df}, with one row per R call. There will be three columns,
\itemize{
\item \code{file}: the path of the original R file
\item \code{expr}: the R call
\item \code{line}: the line of the R call
}
}
\description{
Read R file(s) as a tidy data frame
}
\examples{
d <- read_rfiles(
  tidycode_example("example_plot.R"),
  tidycode_example("example_analysis.R")
)
}
