% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycode_example.R
\name{tidycode_example}
\alias{tidycode_example}
\title{Get path to example file}
\usage{
tidycode_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
tidycode comes bundled with a few small files to use in examples. This
function makes them easy to access.
}
\examples{
tidycode_example()
tidycode_example("example_plot.R")
}
