% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_call.R
\name{unnest_call}
\alias{unnest_call}
\title{Unnest an R call}
\usage{
unnest_call(x, remove_assign = FALSE)
}
\arguments{
\item{x}{an R call}

\item{remove_assign}{logical, indicating whether to remove the assignment
portion of the call, default = `FALSE`.}
}
\value{
a list of the unnested R call
}
\description{
Unnest an R call
}
\examples{
matahari::dance_start()
m <- lm(mpg ~ cyl, mtcars)
matahari::dance_stop()
expr <- matahari::dance_tbl()$expr
unnest_call(expr[[3]])
matahari::dance_remove()

}
