% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_tbl.R
\name{get_packages_tbl}
\alias{get_packages_tbl}
\title{Create a tibble of all functions included in all packages used in your analysis}
\usage{
get_packages_tbl(pkg_names)
}
\arguments{
\item{pkg_names}{Character vector of the names of all packages used in analysis}
}
\value{
Tibble with two columns, `package` and `func`
}
\description{
Create a tibble of all functions included in all packages used in your analysis
}
