% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_model.R
\name{is_model}
\alias{is_model}
\title{Check if an expression is calling a model function}
\usage{
is_model(x)
}
\arguments{
\item{x}{an R expression}
}
\value{
logical
}
\description{
Check if an expression is calling a model function
}
\examples{
matahari::dance_start()
m <- lm(mpg ~ cyl, mtcars)
matahari::dance_stop()
expr <- matahari::dance_tbl()$expr
purrr::map_lgl(expr, is_model)
matahari::dance_remove()
}
