% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_package.R
\name{is_package}
\alias{is_package}
\title{Check if an expression is loading an R package}
\usage{
is_package(x)
}
\arguments{
\item{x}{an R expression}
}
\value{
logical
}
\description{
Check if an expression is loading an R package
}
\examples{
matahari::dance_start()
library(tidycode)
matahari::dance_stop()
expr <- matahari::dance_tbl()$expr
purrr::map_lgl(expr, is_package)
}
