% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_plot.R
\name{is_plot}
\alias{is_plot}
\title{Check if an expression is calling a plot function}
\usage{
is_plot(x)
}
\arguments{
\item{x}{an R expression}
}
\value{
logical
}
\description{
Check if an expression is calling a plot function
}
\examples{
matahari::dance_start()
plot(1:10)
matahari::dance_stop()
expr <- matahari::dance_tbl()$expr
purrr::map_lgl(expr, is_plot)
}
