% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_fun_calls.R
\name{ls_fun_calls}
\alias{ls_fun_calls}
\title{List functions in R call}
\usage{
ls_fun_calls(x)
}
\arguments{
\item{x}{an R call}
}
\value{
Character, all functions in an R call
}
\description{
List functions in R call
}
\examples{
matahari::dance_start()
m <- lm(mpg ~ cyl, mtcars)
matahari::dance_stop()
expr <- matahari::dance_tbl()$expr
ls_fun_calls(expr[[3]])
matahari::dance_remove()
}
