% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_packages.R
\name{ls_packages}
\alias{ls_packages}
\title{List packages}
\usage{
ls_packages(x)
}
\arguments{
\item{x}{an R call or list of R calls}
}
\value{
Character. Vector of packages called.
}
\description{
List packages
}
\examples{
matahari::dance_start()
library(tidycode)
expr <- matahari::dance_tbl()$expr
ls_packages(expr)
matahari::dance_remove()
}
