% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_calls.R
\name{unnest_calls}
\alias{unnest_calls}
\title{Unnest R calls}
\usage{
unnest_calls(x, drop = TRUE)
}
\arguments{
\item{x}{an R call or list of R calls}

\item{drop}{`logical`. Whether the original input R call(s) should be dropped.}
}
\value{
a tibble of the unnested R calls with three columns: `line`: the line number of the call
 (if a list was supplied), `func`: the name of the function called, `args`: a list of arguments
}
\description{
Unnest R calls
}
\examples{
# Unnest a model call
unnest_calls(quote(lm(mpg ~ cyl, mtcars)))

# Unnest a model call and keep the call itself using the drop parameter
unnest_calls(quote(lm(mpg ~ cyl, mtcars)), drop = FALSE)
}
