% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_calls.R
\name{unnest_calls}
\alias{unnest_calls}
\title{Unnest R calls}
\usage{
unnest_calls(x)
}
\arguments{
\item{x}{an R call or list of R calls}
}
\value{
a tibble of the unnested R calls with three columns: `line`: the line number of the call
 (if a list was supplied), `func`: the name of the function called, `args`: a list of arguments
}
\description{
Unnest R calls
}
\examples{
matahari::dance_start()
m <- lm(mpg ~ cyl, mtcars)
matahari::dance_stop()
expr <- matahari::dance_tbl()$expr
unnest_calls(expr)
matahari::dance_remove()
}
